Refactor voq_helpers.py
